Remove durationMs
Actually now that I think about it, tracking 'Duration' does not make that much sense for this game since all leaderboards close anyway after a certain deadline. So it might be better to just keep it for the Training campaign but the change is not worth it at the moment.